Fix bug that made screen look like an animation from start position to current
This was because an eventlistener which is responsible for resetting
chess board UI after game restarts remained active even after restarting.
This meant that whenever a click was registered on the window, UI
reseted to start position until a move was made which set the board to the
current position. Due to this, the Ui kept on reseting for a fration of
second before the player's move and computer's move was shown on the
screen. This resulted in a dragging UI.